requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==23.1.0
    # via
    #   outcome
    #   trio
certifi==2023.5.7
    # via
    #   requests
    #   selenium
charset-normalizer==3.2.0
    # via requests
exceptiongroup==1.1.2
    # via trio-websocket
h11==0.14.0
    # via wsproto
idna==3.4
    # via
    #   requests
    #   trio
outcome==1.2.0
    # via trio
pysocks==1.7.1
    # via urllib3
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
schedule==1.2.0
    # via -r requirements.in
selenium==4.10.0
    # via -r requirements.in
sniffio==1.3.0
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.22.1
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.10.3
    # via selenium
urllib3[socks]==2.0.3
    # via
    #   requests
    #   selenium
wsproto==1.2.0
    # via trio-websocket
